openai/chatgpt-basico/pyproject.toml
[tool.poetry]
name = "chatgpt-basico"
version = "0.1.0"
description = ""
authors = ["Flávio Félix <fau_felix30@hotmail.com>"]
readme = "README.md"
package-mode = false



[tool.poetry.dependencies]
python = "^3.13"
openai = "^2.8.1"
python-dotenv = "^1.2.1"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
